moved FAQInfo, ProfilesCells, SettingsCell, UserInfo to their own files. Formatted and cleaned code in ProfileTab files
