+++ Capstone/Models/MainTabs/ProfileTab/FAQViewController.swift
@@ -8,34 +8,20 @@
 
 import UIKit
 
-struct FAQInfo {
-    let title: String
-    let description: String
-    
-    static func loadFAQs() -> [FAQInfo] {
-        return [
-            FAQInfo(title: "How am I supposed to use the job history feature?", description: "Job History is meant to a \"master resume\" of sorts by allowing you to keep track of your entire job history.\nYou can use your logged job history as reference when creating new resumes, applying to jobs, or simply keeping track of your career over time."),
-            FAQInfo(title: "What is a STAR story?", description: "A STAR story is an example of how you should be answering behavioral interview questions. This method allows you to deliver a clear and impactful response to an interview question.\nThe general structure of the answer should follow: Situation, Task, Action, and Result."),
-            FAQInfo(title: "Where does all of this info go?", description: "We keep a private, secure server on firebase for user data."),
-            FAQInfo(title: "What if I don't want this info uploaded?", description: "We are working on an update to allow our users to store all of their data locally on their device.")
-        ]
-    }
-}
-
 class FAQViewController: UIViewController {
-
+    //MARK:- IBOutlets
     @IBOutlet weak var collectionView: UICollectionView!
     @IBOutlet weak var collectionViewLayout: UICollectionViewFlowLayout!
-    
-    
+    //MARK:- Variables
     private var FAQs = [FAQInfo]()
-    
+    //MARK:- ViewLifeCycles
     override func viewDidLoad() {
         super.viewDidLoad()
         configureNavBar()
         configureCollectionView()
         loadFAQInfo()
     }
+    //MARK:- Functions
     private func configureNavBar() {
         navigationItem.title = "FAQs"
         navigationController?.navigationBar.prefersLargeTitles = true
@@ -44,9 +30,8 @@ class FAQViewController: UIViewController {
         collectionView.delegate = self
         collectionView.dataSource = self
         collectionView.register(UINib(nibName: "BasicInfoCollectionViewCell", bundle: nil), forCellWithReuseIdentifier: "basicInfoCell")
-        
         if let flowLayout = collectionViewLayout,
-           let collectionView = collectionView {
+            let collectionView = collectionView {
             let w = collectionView.frame.width - 20
             flowLayout.estimatedItemSize = CGSize(width: w, height: 200)
         }
@@ -55,7 +40,7 @@ class FAQViewController: UIViewController {
         FAQs = FAQInfo.loadFAQs()
     }
 }
-
+//MARK:- Extensions
 extension FAQViewController: UICollectionViewDelegateFlowLayout {
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
         let faq = FAQs[indexPath.row]
@@ -65,7 +50,6 @@ extension FAQViewController: UICollectionViewDelegateFlowLayout {
         let maxWidth = collectionView.frame.width
         let maxHeight = collectionView.frame.height
         return CGSize(width: maxWidth * 0.90, height: maxHeight / 4)
-//        return CGSize(width: maxWidth * 0.90, height: label.frame.height)
     }
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
         return UIEdgeInsets(top: 10, left: 0, bottom: 10, right: 0)
@@ -75,7 +59,6 @@ extension FAQViewController: UICollectionViewDataSource {
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         return FAQs.count
     }
-    
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
         guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "basicInfoCell", for: indexPath) as? BasicInfoCollectionViewCell else {
             fatalError("Failed to dequeue basicInfoCell")
@@ -86,8 +69,4 @@ extension FAQViewController: UICollectionViewDataSource {
         cell.configureCell(faq: faq, userInfo: nil)
         return cell
     }
-    
-    
-    
-    
 }
