+++ Capstone/Models/MainTabs/ProfileTab/ProfileCell.swift
@@ -9,15 +9,12 @@
 import UIKit
 
 class ProfileCell: UITableViewCell {
-
     @IBOutlet weak var settingsImage: UIImageView!
     @IBOutlet weak var settingsName: UILabel!
     @IBOutlet weak var iconImageView: UIImageView!
-    
     public func configureCell(setting: ProfileCells) {
         settingsImage.image = setting.images
         settingsName.text = setting.title
         iconImageView.tintColor = AppColors.primaryPurpleColor
     }
-    
 }
